+++ cli/php/src/EditCommand.php
@@ -2,6 +2,7 @@
 
 namespace Laraboot;
 
+use Laraboot\Exp\EnvOrDefaultExp;
 use Laraboot\Schema\VisitorContext;
 use Symfony\Component\Console\Command\Command;
 use Laraboot\Schema\PathDefinition;
@@ -36,5 +37,37 @@ class EditCommand extends Command
         }
 
         return new VisitorContext($pathDef, $inputContext);
+    }
+
+    /**
+     * @param array|null $values
+     * @return array
+     */
+    protected function getEnvOrDefaultExps(?array $values): array
+    {
+        return array_map(function ($el) {
+            $orEnv = null;
+            list($key, $value) = explode('=', $el);
+//            echo $value . "\n";
+            if (stripos($value, '|') !== FALSE) {
+                $sub = explode('|', $value);
+                $count = count($sub);
+                if ($count > 1) {
+                    // we have a list of envs
+                    $orEnv = array_splice($sub, 0, $count - 1);
+//                    print_r($orEnv);
+                    $value = array_pop($sub);
+//                    print_r($value);
+                } else {
+                    $orEnv = $sub[0];
+                    $value = $sub[1];
+                }
+            }
+            return new EnvOrDefaultExp([
+                'key' => $key,
+                'value' => $value,
+                'orenv' => $orEnv
+            ]);
+        }, $values);
     }
 }
